Prep for string inputs
