+++ diffusion/models/t5_diffusion.py
@@ -4,23 +4,17 @@
 """Diffusion models."""
 
 import math
-from typing import List, Optional, Tuple, Union
+from typing import List, Optional, Tuple
 
 import torch
 import torch.nn as nn
 import torch.nn.functional as F
-from composer.devices import DeviceGPU
 from composer.models import ComposerModel
 from composer.utils import dist
-from diffusers import AutoencoderKL, DDIMScheduler, DDPMScheduler, EulerDiscreteScheduler, UNet2DConditionModel
 from scipy.stats import qmc
 from torchmetrics import MeanSquaredError
 from tqdm.auto import tqdm
-from transformers import PretrainedConfig
-
-from diffusion.models.autoencoder import AutoEncoder, load_autoencoder
-from diffusion.models.layers import zero_module
-from diffusion.models.models import _parse_latent_statistics
+from transformers import PreTrainedTokenizer
 
 try:
     import xformers  # type: ignore
@@ -77,10 +71,10 @@ class DiffusionV1(ComposerModel):
         vae,
         noise_scheduler,
         inference_noise_scheduler,
-        t5_tokenizer: Optional = None,
-        t5_encoder: Optional = None,
-        clip_tokenizer: Optional = None,
-        clip_encoder: Optional = None,
+        t5_tokenizer: Optional[PreTrainedTokenizer] = None,
+        t5_encoder: Optional[torch.nn.Module] = None,
+        clip_tokenizer: Optional[PreTrainedTokenizer] = None,
+        clip_encoder: Optional[torch.nn.Module] = None,
         prediction_type: str = 'epsilon',
         latent_mean: Tuple[float] = (0.0,) * 4,
         latent_std: Tuple[float] = (1 / 0.13025,) * 4,
@@ -110,8 +104,6 @@ class DiffusionV1(ComposerModel):
         self.quasirandomness = quasirandomness
         self.train_seed = train_seed
         self.val_seed = val_seed
-        self.latent_mean = latent_mean
-        self.latent_std = latent_std
         self.latent_mean = torch.tensor(latent_mean).view(1, -1, 1, 1)
         self.latent_std = torch.tensor(latent_std).view(1, -1, 1, 1)
         self.train_metrics = train_metrics if train_metrics is not None else [MeanSquaredError()]
@@ -148,7 +140,7 @@ class DiffusionV1(ComposerModel):
         return self
 
     def _generate_timesteps(self, latents: torch.Tensor):
-        if not self.model.training:
+        if not self.unet.training:
             # Sample equally spaced timesteps across all devices
             global_batch_size = latents.shape[0] * dist.get_world_size()
             global_timesteps = torch.linspace(0, len(self.noise_scheduler), global_batch_size)
@@ -167,9 +159,9 @@ class DiffusionV1(ComposerModel):
                 timesteps = timesteps[idx_offset:idx_offset + latents.shape[0]].to(latents.device)
             else:
                 timesteps = torch.randint(0,
-                                        len(self.noise_scheduler), (latents.shape[0],),
-                                        device=latents.device,
-                                        generator=self.rng_generator)
+                                          len(self.noise_scheduler), (latents.shape[0],),
+                                          device=latents.device,
+                                          generator=self.rng_generator)
         return timesteps
 
     def set_rng_generator(self, rng_generator: torch.Generator):
@@ -187,7 +179,34 @@ class DiffusionV1(ComposerModel):
         image = self.vae.decode(latents).sample
         image = (image / 2 + 0.5).clamp(0, 1)
         return image
-        
+
+    def encode_text(self, text, device):
+        assert self.t5_tokenizer is not None and self.t5_encoder is not None
+        assert self.clip_tokenizer is not None and self.clip_encoder is not None
+        # Encode with T5
+        t5_tokenizer_out = self.t5_tokenizer(text,
+                                             padding='max_length',
+                                             max_length=self.t5_tokenizer.model_max_length,
+                                             truncation=True,
+                                             return_tensors='pt')
+        tokenized_captions = t5_tokenizer_out['input_ids'].to(device)
+        t5_attn_mask = t5_tokenizer_out['attention_mask'].to(torch.bool).to(device)
+        t5_embed = self.t5_encoder(input_ids=tokenized_captions, attention_mask=t5_attn_mask)
+        # Encode with CLIP
+        clip_tokenizer_out = self.clip_tokenizer(text,
+                                                 padding='max_length',
+                                                 max_length=self.clip_tokenizer.model_max_length,
+                                                 truncation=True,
+                                                 return_tensors='pt')
+        tokenized_captions = clip_tokenizer_out['input_ids'].to(device)
+        clip_attn_mask = clip_tokenizer_out['attention_mask'].to(torch.bool).to(device)
+        clip_out = self.clip_encoder(input_ids=tokenized_captions,
+                                     attention_mask=clip_attn_mask,
+                                     output_hidden_states=True)
+        clip_embed = clip_out.hidden_states[-2]
+        pooled_embeddings = clip_out[1]
+        return t5_embed, clip_embed, t5_attn_mask, clip_attn_mask, pooled_embeddings
+
     def prepare_text_embeddings(self, t5_embed, clip_embed, t5_mask, clip_mask):
         if t5_embed.shape[1] > self.max_seq_len:
             t5_embed = t5_embed[:, :self.max_seq_len]
@@ -278,9 +297,11 @@ class DiffusionV1(ComposerModel):
     @torch.no_grad()
     def generate(
         self,
-        prompt_embeds: torch.FloatTensor,
-        pooled_prompt: torch.FloatTensor,
-        prompt_mask: torch.LongTensor,
+        prompt: Optional[list] = None,
+        negative_prompt: Optional[list] = None,
+        prompt_embeds: Optional[torch.FloatTensor] = None,
+        pooled_prompt: Optional[torch.FloatTensor] = None,
+        prompt_mask: Optional[torch.LongTensor] = None,
         neg_prompt_embeds: Optional[torch.FloatTensor] = None,
         pooled_neg_prompt: Optional[torch.FloatTensor] = None,
         neg_prompt_mask: Optional[torch.LongTensor] = None,
@@ -301,20 +322,11 @@ class DiffusionV1(ComposerModel):
         one forward pass through the unet.
 
         Args:
-            prompt (str or List[str]): The prompt or prompts to guide the image generation.
+            prompt (List[str]): The prompts to guide the image generation.
             negative_prompt (str or List[str]): The prompt or prompts to guide the
                 image generation away from. Ignored when not using guidance
                 (i.e., ignored if guidance_scale is less than 1).
                 Must be the same length as list of prompts. Default: `None`.
-            tokenized_prompts (torch.LongTensor): Optionally pass pre-tokenized prompts instead
-                of string prompts. If SDXL, this will be a tensor of size [B, 2, max_length],
-                otherwise will be of shape [B, max_length]. Default: `None`.
-            tokenized_negative_prompts (torch.LongTensor): Optionally pass pre-tokenized negative
-                prompts instead of string prompts. Default: `None`.
-            tokenized_prompts_pad_mask (torch.LongTensor): Optionally pass padding mask for
-                pre-tokenized prompts. Default `None`.
-            tokenized_negative_prompts_pad_mask (torch.LongTensor): Optionall pass padding mask for
-                pre-tokenized negative prompts. Default `None`.
             prompt_embeds (torch.FloatTensor): Optionally pass pre-tokenized prompts instead
                 of string prompts. If both prompt and prompt_embeds
                 are passed, prompt_embeds will be used. Default: `None`.
@@ -348,17 +360,6 @@ class DiffusionV1(ComposerModel):
                 (representing original size of input image) to use when generating images with SDXL.
                 Default: `None`.
         """
-        # TODO: do checks
-        # if prompt_embeds.shape[:2] == prompt_mask.shape[:2]:
-        #     raise ValueError(' ')
-
-        # Check all parts of negative prompts exist and are equal length
-        # if neg_prompt_embeds is not None or neg_prompt_mask is not None or pooled_neg_prompt is not None:
-
-        # if negative_negative_embedlen(prompt_embeds) != len(negative_prompt_embeds):
-        #     raise ValueError('len(prompts) and len(negative_prompts) must be the same. \
-        #             A negative prompt must be provided for each given prompt.')
-
         # Create rng for the generation
         device = self.vae.device
         rng_generator = torch.Generator(device=device)
@@ -371,6 +372,34 @@ class DiffusionV1(ComposerModel):
             width = self.unet.config.sample_size * self.downsample_factor
 
         do_classifier_free_guidance = guidance_scale > 1.0  # type: ignore
+
+        # Check that inputs are consistent with all embeddings or text inputs. All embeddings should be provided if using
+        # embeddings, and none if using text.
+        if (prompt_embeds is None) == (prompt is None):
+            raise ValueError('One and only one of prompt or prompt_embeds should be provided.')
+        if (pooled_prompt is None) != (prompt_embeds is None):
+            raise ValueError('pooled_prompt should be provided if and only if using embeddings')
+        if (prompt_mask is None) != (prompt_embeds is None):
+            raise ValueError('prompt_mask should be provided if and only if using embeddings')
+        if (neg_prompt_embeds is None) == (negative_prompt is None):
+            raise ValueError('One and only one of negative_prompt or neg_prompt_embeds should be provided.')
+        if (neg_prompt_mask is None) != (neg_prompt_embeds is None):
+            raise ValueError('neg_prompt_mask should be provided if and only if using embeddings')
+        if (pooled_neg_prompt is None) != (neg_prompt_embeds is None):
+            raise ValueError('pooled_neg_prompt should be provided if and only if using embeddings')
+
+        # If the prompt is specified as text, encode it.
+        if prompt is not None:
+            t5_embed, clip_embed, t5_attn_mask, clip_attn_mask, pooled_prompt = self.encode_text(
+                prompt, self.vae.device)
+            prompt_embeds, prompt_mask = self.prepare_text_embeddings(t5_embed, clip_embed, t5_attn_mask,
+                                                                      clip_attn_mask)
+        # If negative prompt is specified as text, encode it.
+        if negative_prompt is not None:
+            t5_embed, clip_embed, t5_attn_mask, clip_attn_mask, pooled_neg_prompt = self.encode_text(
+                negative_prompt, self.vae.device)
+            neg_prompt_embeds, neg_prompt_mask = self.prepare_text_embeddings(t5_embed, clip_embed, t5_attn_mask,
+                                                                              clip_attn_mask)
 
         text_embeddings = _duplicate_tensor(prompt_embeds, num_images_per_prompt)
         pooled_embeddings = _duplicate_tensor(pooled_prompt, num_images_per_prompt)
@@ -468,206 +497,3 @@ def _duplicate_tensor(tensor, num_images_per_prompt):
     return tensor.view(batch_size * num_images_per_prompt, seq_len, *[
         -1,
     ] * len(tensor.shape[2:]))
-
-
-def build_diffusion_v1(
-    unet_model_name: str = 'stabilityai/stable-diffusion-xl-base-1.0',
-    vae_model_name: str = 'madebyollin/sdxl-vae-fp16-fix',
-    autoencoder_path: Optional[str] = None,
-    autoencoder_local_path: str = '/tmp/autoencoder_weights.pt',
-    prediction_type: str = 'epsilon',
-    latent_mean: Union[float, Tuple, str] = 0.0,
-    latent_std: Union[float, Tuple, str] = 7.67754318618,
-    text_embed_dim: int = 4096,
-    beta_schedule: str = 'scaled_linear',
-    zero_terminal_snr: bool = False,
-    train_metrics: Optional[List] = None,
-    val_metrics: Optional[List] = None,
-    quasirandomness: bool = False,
-    train_seed: int = 42,
-    val_seed: int = 1138,
-    fsdp: bool = True,
-    use_xformers: bool = True,
-):
-    """Stable diffusion 2 training setup + SDXL UNet and VAE.
-
-    Requires batches of matched images and text prompts to train. Generates images from text
-    prompts. Currently uses UNet and VAE config from SDXL, but text encoder/tokenizer from SD2.
-
-    Args:
-        unet_model_name (str): Name of the UNet model to load. Defaults to
-            'stabilityai/stable-diffusion-xl-base-1.0'.
-        vae_model_name (str): Name of the VAE model to load. Defaults to
-            'madebyollin/sdxl-vae-fp16-fix' as the official VAE checkpoint (from
-            'stabilityai/stable-diffusion-xl-base-1.0') is not compatible with fp16.
-        autoencoder_path (optional, str): Path to autoencoder weights if using custom autoencoder. If not specified,
-            will use the vae from `model_name`. Default `None`.
-        autoencoder_local_path (optional, str): Path to autoencoder weights. Default: `/tmp/autoencoder_weights.pt`.
-        prediction_type (str): The type of prediction to use. Must be one of 'sample',
-            'epsilon', or 'v_prediction'. Default: `epsilon`.
-        latent_mean (float, Tuple, str): The mean of the autoencoder latents. Either a float for a single value,
-            a tuple of means, or or `'latent_statistics'` to try to use the value from the autoencoder
-            checkpoint. Defaults to `0.0`.
-        latent_std (float, Tuple, str): The std. dev. of the autoencoder latents. Either a float for a single value,
-            a tuple of std_devs, or or `'latent_statistics'` to try to use the value from the autoencoder
-            checkpoint. Defaults to `1/0.13025`.
-        beta_schedule (str): The beta schedule to use. Must be one of 'scaled_linear', 'linear', or 'squaredcos_cap_v2'.
-            Default: `scaled_linear`.
-        zero_terminal_snr (bool): Whether to enforce zero terminal SNR. Default: `False`.
-        train_metrics (list, optional): List of metrics to compute during training. If None, defaults to
-            [MeanSquaredError()].
-        val_metrics (list, optional): List of metrics to compute during validation. If None, defaults to
-            [MeanSquaredError()].
-        quasirandomness (bool): Whether to use quasirandomness for generating diffusion process noise.
-            Default: `False`.
-        train_seed (int): Seed to use for generating diffusion process noise during training if using
-            quasirandomness. Default: `42`.
-        val_seed (int): Seed to use for generating evaluation images. Defaults to 1138.
-        fsdp (bool): Whether to use FSDP. Defaults to True.
-        use_xformers (bool): Whether to use xformers for attention. Defaults to True.
-    """
-    latent_mean, latent_std = _parse_latent_statistics(latent_mean), _parse_latent_statistics(latent_std)
-
-    if train_metrics is None:
-        train_metrics = [MeanSquaredError()]
-    if val_metrics is None:
-        val_metrics = [MeanSquaredError()]
-
-    # Make the autoencoder
-    if autoencoder_path is None:
-        if latent_mean == 'latent_statistics' or latent_std == 'latent_statistics':
-            raise ValueError('Cannot use tracked latent_statistics when using the pretrained vae.')
-        downsample_factor = 8
-        # Use the pretrained vae
-        try:
-            vae = AutoencoderKL.from_pretrained(vae_model_name, subfolder='vae', torch_dtype=torch.float16)
-        except:  # for handling SDXL vae fp16 fixed checkpoint
-            vae = AutoencoderKL.from_pretrained(vae_model_name, torch_dtype=torch.float16)
-    else:
-        # Use a custom autoencoder
-        vae, latent_statistics = load_autoencoder(autoencoder_path, autoencoder_local_path, torch_dtype=torch.float16)
-        if latent_statistics is None and (latent_mean == 'latent_statistics' or latent_std == 'latent_statistics'):
-            raise ValueError(
-                'Must specify latent scale when using a custom autoencoder without tracking latent statistics.')
-        if isinstance(latent_mean, str) and latent_mean == 'latent_statistics':
-            assert isinstance(latent_statistics, dict)
-            latent_mean = tuple(latent_statistics['latent_channel_means'])
-        if isinstance(latent_std, str) and latent_std == 'latent_statistics':
-            assert isinstance(latent_statistics, dict)
-            latent_std = tuple(latent_statistics['latent_channel_stds'])
-        downsample_factor = 2**(len(vae.config['channel_multipliers']) - 1)
-
-    # Make the unet
-    unet_config = PretrainedConfig.get_config_dict(unet_model_name, subfolder='unet')[0]
-
-    if isinstance(vae, AutoEncoder):
-        # Adapt the unet config to account for differing number of latent channels if necessary
-        unet_config['in_channels'] = vae.config['latent_channels']
-        unet_config['out_channels'] = vae.config['latent_channels']
-    unet_config['cross_attention_dim'] = text_embed_dim
-    # This config variable is the sum of the text encoder projection dimension (768 for CLIP) and
-    # the number of additional time embeddings (6) * addition_time_embed_dim (256)
-    unet_config['projection_class_embeddings_input_dim'] = 2304
-    # Init the unet from the config
-    unet = UNet2DConditionModel(**unet_config)
-
-    # Zero initialization trick
-    for name, layer in unet.named_modules():
-        # Final conv in ResNet blocks
-        if name.endswith('conv2'):
-            layer = zero_module(layer)
-        # proj_out in attention blocks
-        if name.endswith('to_out.0'):
-            layer = zero_module(layer)
-    # Last conv block out projection
-    unet.conv_out = zero_module(unet.conv_out)
-
-    if isinstance(latent_mean, float):
-        latent_mean = (latent_mean,) * unet_config['in_channels']
-    if isinstance(latent_std, float):
-        latent_std = (latent_std,) * unet_config['in_channels']
-    assert isinstance(latent_mean, tuple) and isinstance(latent_std, tuple)
-
-    # FSDP Wrapping Scheme
-    if hasattr(unet, 'mid_block') and unet.mid_block is not None:
-        for attention in unet.mid_block.attentions:
-            attention._fsdp_wrap = True
-        for resnet in unet.mid_block.resnets:
-            resnet._fsdp_wrap = True
-    for block in unet.up_blocks:
-        if hasattr(block, 'attentions'):
-            for attention in block.attentions:
-                attention._fsdp_wrap = True
-        if hasattr(block, 'resnets'):
-            for resnet in block.resnets:
-                resnet._fsdp_wrap = True
-    for block in unet.down_blocks:
-        if hasattr(block, 'attentions'):
-            for attention in block.attentions:
-                attention._fsdp_wrap = True
-        if hasattr(block, 'resnets'):
-            for resnet in block.resnets:
-                resnet._fsdp_wrap = True
-
-    # Make the noise schedulers
-    noise_scheduler = DDPMScheduler(num_train_timesteps=1000,
-                                    beta_start=0.00085,
-                                    beta_end=0.012,
-                                    beta_schedule=beta_schedule,
-                                    trained_betas=None,
-                                    variance_type='fixed_small',
-                                    clip_sample=False,
-                                    prediction_type=prediction_type,
-                                    sample_max_value=1.0,
-                                    timestep_spacing='leading',
-                                    steps_offset=1,
-                                    rescale_betas_zero_snr=zero_terminal_snr)
-    if beta_schedule == 'squaredcos_cap_v2':
-        inference_noise_scheduler = DDIMScheduler(num_train_timesteps=1000,
-                                                  beta_start=0.00085,
-                                                  beta_end=0.012,
-                                                  beta_schedule=beta_schedule,
-                                                  trained_betas=None,
-                                                  clip_sample=False,
-                                                  set_alpha_to_one=False,
-                                                  prediction_type=prediction_type,
-                                                  rescale_betas_zero_snr=zero_terminal_snr)
-    else:
-        inference_noise_scheduler = EulerDiscreteScheduler(num_train_timesteps=1000,
-                                                           beta_start=0.00085,
-                                                           beta_end=0.012,
-                                                           beta_schedule=beta_schedule,
-                                                           trained_betas=None,
-                                                           prediction_type=prediction_type,
-                                                           interpolation_type='linear',
-                                                           use_karras_sigmas=False,
-                                                           timestep_spacing='leading',
-                                                           steps_offset=1,
-                                                           rescale_betas_zero_snr=zero_terminal_snr)
-
-    # Make the composer model
-    model = DiffusionV1(
-        unet=unet,
-        vae=vae,
-        noise_scheduler=noise_scheduler,
-        inference_noise_scheduler=inference_noise_scheduler,
-        prediction_type=prediction_type,
-        latent_mean=latent_mean,
-        latent_std=latent_std,
-        downsample_factor=downsample_factor,
-        train_metrics=train_metrics,
-        val_metrics=val_metrics,
-        quasirandomness=quasirandomness,
-        train_seed=train_seed,
-        val_seed=val_seed,
-        text_embed_dim=text_embed_dim,
-        fsdp=fsdp,
-    )
-    if torch.cuda.is_available():
-        model = DeviceGPU().module_to_device(model)
-        if is_xformers_installed and use_xformers:
-            model.unet.enable_xformers_memory_efficient_attention()
-            if hasattr(model.vae, 'enable_xformers_memory_efficient_attention'):
-                model.vae.enable_xformers_memory_efficient_attention()
-
-    return model
